+++ app/src/main/java/com/inky/fitnesscalendar/repository/backup/BackupRepository.kt
@@ -3,6 +3,7 @@ package com.inky.fitnesscalendar.repository.backup
 import android.content.Context
 import android.net.Uri
 import android.os.Build
+import android.util.Log
 import androidx.annotation.StringRes
 import androidx.compose.runtime.Immutable
 import androidx.core.net.toUri
@@ -11,19 +12,23 @@ import com.inky.fitnesscalendar.R
 import com.inky.fitnesscalendar.db.AppDatabase
 import com.inky.fitnesscalendar.util.BACKUP_CACHE_FILE
 import com.inky.fitnesscalendar.util.SDK_MIN_VERSION_FOR_SQLITE_VACUUM
-import com.inky.fitnesscalendar.util.Zip
+import com.inky.fitnesscalendar.util.ZipWriter
 import com.inky.fitnesscalendar.util.copyFile
+import com.inky.fitnesscalendar.util.restartApplication
 import com.inky.fitnesscalendar.util.toLocalDateTime
 import dagger.hilt.android.qualifiers.ApplicationContext
 import java.io.File
+import java.io.IOException
 import java.time.Instant
 import java.time.LocalDateTime
 import java.util.Date
+import java.util.zip.ZipInputStream
 import javax.inject.Inject
 import javax.inject.Singleton
 
-const val BACKUP_IMAGES_PATH = "images"
-const val BACKUP_NAME = "backup.zip"
+private const val BACKUP_NAME = "backup.zip"
+
+private const val TAG = "BackupRepository"
 
 @Immutable
 @Singleton
@@ -34,6 +39,10 @@ class BackupRepository @Inject constructor(
     enum class BackupError(@StringRes val msgID: Int) {
         OldAndroidVersion(R.string.your_android_version_is_too_old_for_backup),
         CannotAccessFile(R.string.cannot_access_file)
+    }
+
+    enum class RestoreError(@StringRes val msgID: Int) {
+        IOError(R.string.cannot_read_file)
     }
 
     fun getLastBackup(directory: Uri): LocalDateTime? {
@@ -56,7 +65,7 @@ class BackupRepository @Inject constructor(
 
         val zipFile = File(context.cacheDir, BACKUP_CACHE_FILE)
 
-        Zip(zipFile).use { zip ->
+        ZipWriter(zipFile).use { zip ->
             for (location in BACKUP_LOCATIONS) {
                 location.backup(context, database, zip)
             }
@@ -65,6 +74,25 @@ class BackupRepository @Inject constructor(
         context.copyFile(zipFile.toUri(), targetUri)
 
         return null
+    }
+
+    /**
+     * Restores a backup from the given zip file and restarts the app, or returns an error
+     */
+    fun restore(uri: Uri): RestoreError {
+        try {
+            context.contentResolver.openInputStream(uri).use { inputStream ->
+                val zip = ZipInputStream(inputStream)
+                for (location in BACKUP_LOCATIONS) {
+                    location.restore(context, database, zip)
+                }
+            }
+        } catch (e: IOException) {
+            Log.e(TAG, "Could not restore: ${e.printStackTrace()}")
+            return RestoreError.IOError
+        }
+
+        restartApplication(context)
     }
 
     /**