+++ app/src/main/java/com/inky/fitnesscalendar/util/ZipWriter.kt
@@ -8,13 +8,14 @@ import java.util.zip.ZipEntry
 import java.util.zip.ZipOutputStream
 import kotlin.io.path.name
 
-class Zip(output: File) : Closeable {
+class ZipWriter(output: File) : Closeable {
     private val zipStream = ZipOutputStream(FileOutputStream(output))
 
     fun addFile(file: File, directory: String = "") {
         assert(file.isFile) { "Recursion should not be required right now" }
+        val filename = (directory + "/" + file.toPath().name).trimStart('/')
         FileInputStream(file).use { inputStream ->
-            zipStream.putNextEntry(ZipEntry(directory + "/" + file.toPath().name))
+            zipStream.putNextEntry(ZipEntry(filename))
             inputStream.copyTo(zipStream)
         }
     }