+++ app/src/main/java/com/inky/fitnesscalendar/util/zip_utils.kt
@@ -1,0 +1,12 @@
+package com.inky.fitnesscalendar.util
+
+import java.util.zip.ZipEntry
+import java.util.zip.ZipInputStream
+
+inline fun ZipInputStream.entries(handler: (ZipEntry) -> Unit) {
+    var entry = nextEntry
+    while (entry != null) {
+        handler(entry)
+        entry = nextEntry
+    }
+}